refactor: move infrastructure index package into io
