+++ src/main/java/pl/koder95/eme/io/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Implementacje I/O i parserów indeksów.
  */
-package pl.koder95.eme.infrastructure.index;
+package pl.koder95.eme.io;
